# use XORBytes utility in crypto/subtle