--- a/go.mod
+++ b/go.mod
@@ -1,8 +1,7 @@
 module github.com/kasperdi/SPHINCSPLUS-golang
 
-go 1.15
+go 1.20
 
-require (
-	golang.org/x/crypto v0.0.0-20220518034528-6f7dac969898
-	golang.org/x/sys v0.0.0-20220520151302-bc2c85ada10a // indirect
-)
+require golang.org/x/crypto v0.0.0-20220518034528-6f7dac969898
+
+require golang.org/x/sys v0.0.0-20220520151302-bc2c85ada10a // indirect