--- a/tweakable/sha256Tweak.go
+++ b/tweakable/sha256Tweak.go
@@ -3,6 +3,7 @@ package tweakable
 import (
 	"crypto/hmac"
 	"crypto/sha256"
+	"crypto/subtle"
 
 	"github.com/kasperdi/SPHINCSPLUS-golang/address"
 	"github.com/kasperdi/SPHINCSPLUS-golang/util"
@@ -51,7 +52,8 @@ func (h *Sha256Tweak) F(PKseed []byte, adrs *address.ADRS, tmp []byte) []byte {
 
 	if h.Variant == Robust {
 		bitmask := mgf1sha256(append(PKseed, compressedADRS...), len(tmp))
-		M1 = util.XorBytes(tmp, bitmask)
+		M1 = make([]byte, len(tmp))
+		_ = subtle.XORBytes(M1, tmp, bitmask)
 	} else if h.Variant == Simple {
 		M1 = tmp
 	}